Refactored all JS. Added licenses to about-panel.
