+++ index.html
@@ -96,6 +96,15 @@
         <p>
             This is a a little project I (<a href="https://github.com/BurnyLlama">BurnyLlama</a>) made in school. :D
         </p>
+
+        <h2>Licensing</h2>
+        <p>
+            The font used on this site is <a href="https://fonts.google.com/specimen/Montserrat?selection.family=Montserrat&sidebar.open=true&preview.text_type=custom">Montserrat</a>.
+            Montserrat is available under the <a href="./licenses/ofl.html">SIL OFL</a> license.
+        </p>
+        <p>
+            The rest of the site is licensed under the <a href="./licenses/mit.html">MIT license</a>.
+        </p>
     </div>
 
     <div class="panel panel-right" id="help-panel" direction="right">
@@ -122,9 +131,6 @@
 
     <img id="close-btn" src="./assets/close.svg" alt="x">
 
-    <script src="./assets/search.js"></script>
-    <script src="./assets/searchmsg.js"></script>
-    <script src="./assets/panels.js"></script>
-    <script src="./assets/settings.js"></script>
+    <script src="./scripts/main.js" type="module"></script>
 </body>
 </html>
