+++ licenses/mit.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>MIT</title>
+    <style>
+        body {
+            color: #222;
+            background: #fafafa;
+            display: flex;
+        }
+
+        .wrapper {
+            width: 80vw;
+            margin: 5vh auto;
+        }
+    </style>
+</head>
+<body>
+    <div class="wrapper">
+        <h1>
+            The MIT License (MIT)
+        </h1>
+
+        <h2>
+            Copyright © 2021 Fabian Seluska (BurnyLlama)
+        </h2>
+        
+        <p>
+            Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions: <br><br>
+
+            The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software. <br><br>
+
+            THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+        </p>
+    </div>
+</body>
+</html>
