+++ scripts/main.js
@@ -1,0 +1,7 @@
+import panels from "./panels.js"
+import search from "./search.js"
+import settings from "./settings.js"
+
+search.init()
+panels.init()
+settings.init()
